+++ src/BugsnagUnity/Native/Cocoa/NativeClient.cs
@@ -28,13 +28,14 @@ namespace BugsnagUnity
      */
     IntPtr CreateNativeConfig(IConfiguration config) {
       IntPtr obj = NativeCode.bugsnag_createConfiguration(config.ApiKey);
-      NativeCode.bugsnag_setAutoNotify(obj, config.AutoNotify);
+      NativeCode.bugsnag_setAppHangs(obj, true);
+      NativeCode.bugsnag_setAutoNotifyConfig(obj, config.AutoNotify);
       NativeCode.bugsnag_setReleaseStage(obj, config.ReleaseStage);
       NativeCode.bugsnag_setAppVersion(obj, config.AppVersion);
       NativeCode.bugsnag_setNotifyUrl(obj, config.Endpoint.ToString());
 
       if (config.Context != null) {
-        NativeCode.bugsnag_setContext(obj, config.Context);
+        NativeCode.bugsnag_setContextConfig(obj, config.Context);
       }
       var releaseStages = config.NotifyReleaseStages;
       if (releaseStages != null) {
@@ -177,7 +178,7 @@ namespace BugsnagUnity
 
     public void SetAutoNotify(bool autoNotify)
     {
-      NativeCode.bugsnag_setAutoNotify(NativeConfiguration, autoNotify);
+      NativeCode.bugsnag_setAutoNotify(autoNotify);
     }
 
     public void SetAutoDetectAnrs(bool autoDetectAnrs)