+++ src/BugsnagUnity/Native/Cocoa/NativeCode.cs
@@ -36,10 +36,19 @@ namespace BugsnagUnity
     internal static extern void bugsnag_setReleaseStage(IntPtr configuration, string releaseStage);
 
     [DllImport(Import)]
-    internal static extern void bugsnag_setAutoNotify(IntPtr configuration, bool autoNotify);
+    internal static extern void bugsnag_setAutoNotify(bool autoNotify);
+
+    [DllImport(Import)]
+    internal static extern void bugsnag_setAutoNotifyConfig(IntPtr configuration, bool autoNotify);
+
+    [DllImport(Import)]
+    internal static extern void bugsnag_setAppHangs(IntPtr configuration, bool appHangs);
 
     [DllImport(Import)]
     internal static extern void bugsnag_setContext(IntPtr configuration, string context);
+
+    [DllImport(Import)]
+    internal static extern void bugsnag_setContextConfig(IntPtr configuration, string context);
 
     [DllImport(Import)]
     internal static extern void bugsnag_setAppVersion(IntPtr configuration, string appVersion);