+++ test/mobile/features/fixtures/maze_runner/Assets/Scripts/ReporterBehavior.cs
@@ -14,16 +14,19 @@ public class ReporterBehavior : MonoBehaviour {
         config.SessionEndpoint = new Uri("http://bs-local.com:9339/sessions");
         config.Context = "My context";
         config.AppVersion = "1.2.3";
+        config.NotifyLevel = LogType.Error;
         Bugsnag.Start(config);
+        LeaveBreadcrumbString();
+        LeaveBreadcrumbTuple();
     }
-
+     
     public void ThrowException() {
         throw new System.Exception("You threw an exception!");
     }
 
-    public void AssertionFailure() {
-        var items = new int[]{1, 2, 3};
-        Debug.Log("Item4 is: " + items[4]);
+    public void LogError() {
+        SetUser();
+        Debug.LogError("Something went wrong.");
     }
 
     public void NativeException() {
@@ -39,8 +42,8 @@ public class ReporterBehavior : MonoBehaviour {
         }
     }
 
-    public void LogWithClassPrefix() {
-        throw new ExecutionEngineException("Haven't gotten around to making this work, sorry");
+    public void NdkSignal() {
+        BugsnagNative.RaiseNdkSignal();
     }
 
     public void NotifyCaughtException() {